# frontend/person-blog-app/packages/blog/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gateway.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/video"
    }
  ],
  "paths": {
...
    "/api/Channel/posts/{channelId}": {
      "get": {
        "tags": [
          "Channel"
        ],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCommonModelV2PagedListViewModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Post/detail/{postId}": {
      "get": {
        "tags": [
          "Post"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailViewModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/TextPost/{postId}": {
      "get": {
        "tags": [
          "TextPost"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextPostDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/Video/{blogId}/{postId}/{file}": {
      "get": {
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Video/video/{postId}": {
      "get": {
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fromPlaylist",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/VideoUpload/initiate": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/generate-url": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreSignedUrl"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/complete": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/abort": {
      "post": {
        "tags": [
          "VideoUpload"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AbortUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/VideoUpload/session/{uploadId}": {
      "get": {
        "tags": [
          "VideoUpload"
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultipartUploadSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/VideoUpload/parts/{uploadId}": {
      "get": {
        "tags": [
          "VideoUpload"
        ],
        "parameters": [
          {
            "name": "uploadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MultipartUploadPart"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbortUploadRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CompleteUploadRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultipartUploadPart"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GenerateUrlRequest": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "expiryMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "InitiateUploadRequest": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "format": "uuid"
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "fileExtension": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "MultipartUploadPart": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "eTag": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "MultipartUploadSession": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "bucketId": {
            "type": "string",
            "nullable": true
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/UploadStatus"
          },
          "totalSize": {
            "type": "integer",
            "format": "int64"
          },
          "totalParts": {
            "type": "integer",
            "format": "int32"
          },
          "completedETag": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostCommonModelV2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "postType": {
            "$ref": "#/components/schemas/PostTypeModel"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PostCommonModelV2PagedListViewModel": {
        "required": [
          "items",
          "pageSize",
          "totalPageCount",
          "totalPostsCount"
        ],
        "type": "object",
        "properties": {
          "totalPageCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPostsCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostCommonModelV2"
            }
          }
        },
        "additionalProperties": false
      },
...
      "PostDetailViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/PostType"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "dislikeCount": {
            "type": "integer",
            "format": "int32"
          },
          "videoData": {
            "$ref": "#/components/schemas/VideoMetadataModel"
          },
          "isProcessed": {
            "$ref": "#/components/schemas/ProcessState"
          }
        },
        "additionalProperties": false
      },
...
      "PostType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "PostTypeModel": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "PreSignedUrl": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "httpMethod": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProcessState": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "TextPostAuthor": {
        "type": "object",
        "properties": {
          "blogId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "subscribersCount": {
            "type": "integer",
            "format": "int32"
          },
          "postsCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalViewsCount": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "TextPostCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TextPostDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "lead": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "estimatedReadingTimeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int64"
          },
          "likeCount": {
            "type": "integer",
            "format": "int32"
          },
          "dislikeCount": {
            "type": "integer",
            "format": "int32"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextPostCategory"
            },
            "nullable": true
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextPostMedia"
            },
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/TextPostAuthor"
          },
          "viewer": {
            "$ref": "#/components/schemas/TextPostViewerState"
          }
        },
        "additionalProperties": false
      },
...
      "TextPostMedia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "length": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "TextPostViewerState": {
        "type": "object",
        "properties": {
          "isViewed": {
            "type": "boolean"
          },
          "isLike": {
            "type": "boolean",
            "nullable": true
          },
          "isSubscribed": {
            "type": "boolean"
          },
          "canEdit": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UploadStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "VideoMetadataModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "length": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "objectName": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}